+++ file.py
@@ -33,7 +33,7 @@ autosave_dir_path = os.path.join(temp_dir_path, "hot_node_autosave")
 history_dir_path = os.path.join(addon_dir_path, "hot_node_history")
 pack_root_dir_path = os.path.join(addon_dir_path, "preset_packs")
 pack_selected_dir_path = os.path.join(pack_root_dir_path, "")
-pack_meta_path = ""
+pack_selected_meta_path = ""
 root_meta_path = os.path.join(pack_root_dir_path, ".metadata.json")
 
 # Metas
@@ -108,7 +108,7 @@ def check_preset_existing(preset_name):
 
 
 # Get File Info & Path
-def get_pack_meta_path():
+def get_pack_selected_meta_path():
     os.path.join(pack_selected_dir_path, ".metadata.json")
     
     
@@ -119,6 +119,10 @@ def get_pack_meta_path(pack_name):
 
 def get_pack_path(pack_name):
     return os.path.join(pack_root_dir_path, pack_name)
+
+
+def get_preset_path(preset_name):
+    return os.path.join(pack_selected_dir_path, preset_name)
     
     
 def get_root_meta_path():
@@ -157,7 +161,7 @@ def read_json(file_path) -> dict:
     
 def write_meta():
     update_root_meta_cache_mtime()
-    write_json(pack_meta_path, pack_meta_cache)
+    write_json(pack_selected_meta_path, pack_meta_cache)
     write_json(root_meta_path, root_meta_cache)
     
     
@@ -167,8 +171,8 @@ def write_root_meta():
     
 def read_pack_meta(pack_name: str|None=None):
     if pack_name is None:
-        if os.path.exists(pack_meta_path):
-            return read_json(pack_meta_path)
+        if os.path.exists(pack_selected_meta_path):
+            return read_json(pack_selected_meta_path)
     else:
         meta_path = os.path.join(pack_root_dir_path, pack_name, ".metadata.json")
         if os.path.exists(meta_path):
@@ -181,8 +185,8 @@ def read_root_meta():
     
     
 def read_metas():
-    if os.path.exists(root_meta_path) and os.path.exists(pack_meta_path):
-        return read_json(root_meta_path), read_json(pack_meta_path)
+    if os.path.exists(root_meta_path) and os.path.exists(pack_selected_meta_path):
+        return read_json(root_meta_path), read_json(pack_selected_meta_path)
     
     
 def refresh_root_meta_cache():
@@ -218,19 +222,49 @@ def read_existing_files(dir_path, suffix=".zip", cull_suffix=True):
 
 
 # History Record
-def save_history(src_path):
+def push_history(ori_paths, extra_identifier="d"):
     '''Copy the file to the history folder and return the copied file path.'''
-    name = os.path.basename(src_path)
-    record_time = str(time.time())
-    dot_suffix = utils.get_dot_suffix(src_path, ".zip", ".json")
-    if dot_suffix is None:
-        his_path = os.path.join(history_dir_path, record_time)
-        shutil.copytree(src_path, his_path)
-    else:
-        dst_name = "".join((record_time, dot_suffix))
-        his_path = os.path.join(history_dir_path, dst_name)
-        shutil.copyfile(src_path, his_path)
-    return his_path
+    his_paths = []
+    path_num = len(ori_paths)
+    for i in range(path_num):
+        ori_path = ori_paths[i]
+        identifier = "_".join((str(time.time()), str(i), extra_identifier))
+        dot_suffix = utils.get_dot_suffix(ori_path, ".zip", ".json")
+        if dot_suffix is None:
+            his_path = os.path.join(history_dir_path, identifier)
+            shutil.copytree(ori_path, his_path)
+        else:
+            dst_name = "".join((identifier, dot_suffix))
+            his_path = os.path.join(history_dir_path, dst_name)
+            shutil.copyfile(ori_path, his_path)
+        his_paths.append(his_path)
+    return his_paths
+
+
+def pull_history(ori_paths, his_paths):
+    path_num = len(ori_paths)
+    for i in range(path_num):
+        ori_path = ori_paths[i]
+        his_path = his_paths[i]
+        dot_suffix = utils.get_dot_suffix(ori_path, ".zip", ".json")
+        if dot_suffix is None:
+            if os.path.exists(ori_path):
+                shutil.rmtree(ori_path)
+            shutil.copytree(his_path, ori_path)
+            shutil.rmtree(his_path)
+        else:
+            if os.path.exists(ori_path):
+                os.remove(ori_path)
+            shutil.copyfile(his_path, ori_path)
+            os.remove(his_path)
+            
+    
+def del_paths(paths):
+    for path in paths:
+        if os.path.isdir(path):
+            shutil.rmtree(path)
+        else:
+            os.remove(path)
     
 
 # CRUD of Pack and Preset
@@ -238,9 +272,9 @@ def create_pack(pack_name):
     global root_meta_cache
     global pack_meta_cache
     global pack_selected_dir_path
-    global pack_meta_path
+    global pack_selected_meta_path
     pack_selected_dir_path = os.path.join(pack_root_dir_path, pack_name)
-    pack_meta_path = os.path.join(pack_selected_dir_path, ".metadata.json")
+    pack_selected_meta_path = os.path.join(pack_selected_dir_path, ".metadata.json")
     # create pack metadata
     root_meta_cache["pack_selected"] = pack_name
     pack_meta_cache = {}
@@ -251,12 +285,14 @@ def create_pack(pack_name):
     write_meta()
     # reload packs to get right order of packs
     load_packs_and_get_names()
+    return pack_selected_dir_path
 
 
 def delete_pack(pack_name):
     global pack_meta_cache
     pack_dir_path = os.path.join(pack_root_dir_path, pack_name)
     shutil.rmtree(pack_dir_path)
+    return pack_dir_path
 
 
 def clear_pack(pack_names):
@@ -285,8 +321,7 @@ def load_packs_and_get_names():
     new_pack_num = len(pack_names)
     for i in range(new_pack_num):
         pack_name = pack_names[i]
-        pack_mtime = get_pack_mtime(pack_name)
-        pack = props_py.Pack(pack_name, pack_mtime)
+        pack = props_py.Pack(pack_name)
         props_py.gl_packs[pack_name] = pack
     return pack_names
 
@@ -294,18 +329,18 @@ def load_packs_and_get_names():
 def select_pack(pack: props_py.Pack|None):
     global pack_meta_cache
     global root_meta_cache
-    global pack_meta_path
+    global pack_selected_meta_path
     global pack_selected_dir_path
     if pack is None:
         pack_meta_cache = {}
         pack_meta_cache["order"] = []
         pack_meta_cache["tree_types"] = {}
         pack_meta_cache["version"] = version
-        pack_meta_path = ""
+        pack_selected_meta_path = ""
         root_meta_cache["pack_selected"] = ""
     else:
         pack_selected_dir_path = os.path.join(pack_root_dir_path, pack.name)
-        pack_meta_path = os.path.join(pack_selected_dir_path, ".metadata.json")
+        pack_selected_meta_path = os.path.join(pack_selected_dir_path, ".metadata.json")
         pack_meta_cache = read_pack_meta()
         root_meta_cache["pack_selected"] = pack.name
     props_py.gl_pack_selected = pack
@@ -438,6 +473,7 @@ def create_preset(preset_name: str, cpreset: dict):
     pack_meta_cache["version"] = version
     write_json(file_path, cpreset)
     write_meta()
+    return file_path
     
     
 def update_preset(preset_name: str, cpreset: dict):
@@ -448,6 +484,7 @@ def update_preset(preset_name: str, cpreset: dict):
     pack_meta_cache["version"] = version
     write_json(file_path, cpreset)
     write_meta()
+    return file_path
 
 
 def delete_preset(preset_name):
@@ -458,6 +495,7 @@ def delete_preset(preset_name):
     del pack_meta_cache["tree_types"][preset_name]
     os.remove(file_path)
     write_meta()
+    return file_path
     
     
 def clear_preset(pack_name):